+++ packages/generator/lib/combo/randomizer/checks.ts
@@ -346,7 +346,8 @@ function checkKey(check: WorldCheck): number {
     case 'fairy_spot':
     case 'wonder':
     case 'butterfly':
-    case 'redboulder':
+    case 'boulder-red':
+    case 'boulder-silver':
     case 'icicle':
     case 'redice':
       /* xflag */
@@ -376,7 +377,8 @@ function checkKey(check: WorldCheck): number {
     case 'fairy_spot':
     case 'wonder':
     case 'butterfly':
-    case 'redboulder':
+    case 'boulder-red':
+    case 'boulder-silver':
     case 'icicle':
     case 'redice':
       sceneId = (SCENES as any)[check.scene];